Add question management functionality
This commit adds the capability to manage questions in the application. It includes the creation of a new Question model, repository, service and controller. It also includes necessary changes in the application.properties file and the creation of GraphQL schema. This functionality adds the ability to add, retrieve and delete questions.